+++ test/forgi/threedee/utilities/test_dssr.py
@@ -4,13 +4,13 @@ import forgi.threedee.model.coarse_grain as ftmc
 
 class TestHelperFunctions(unittest.TestCase):
     def test_dssr_to_pdb_atom_id_validIds(self):
-        self.assertEqual(ftud.dssr_to_pdb_atom_id("B.C24"), (" ", 24, " "))
-        self.assertEqual(ftud.dssr_to_pdb_atom_id("LYS124"), (" ", 124, " "))
-        self.assertEqual(ftud.dssr_to_pdb_atom_id("Z12.U13"), (" ", 13, " "))
-        self.assertEqual(ftud.dssr_to_pdb_atom_id("A.5BU36"), (" ", 36, " "))
-        self.assertEqual(ftud.dssr_to_pdb_atom_id("C.C47^M"), (" ", 47, "M"))
-        self.assertEqual(ftud.dssr_to_pdb_atom_id("C.5BU47^M"), (" ", 47, "M"))
-        self.assertEqual(ftud.dssr_to_pdb_atom_id(u'A.C1'), (" ", 1, " "))
+        self.assertEqual(ftud.dssr_to_pdb_atom_id("B.C24"), ("B",(" ", 24, " ")))
+        self.assertEqual(ftud.dssr_to_pdb_atom_id("LYS124"), (None, (" ", 124, " ")))
+        self.assertEqual(ftud.dssr_to_pdb_atom_id("Z12.U13"), ("Z12", (" ", 13, " ")))
+        self.assertEqual(ftud.dssr_to_pdb_atom_id("A.5BU36"), ("A", (" ", 36, " ")))
+        self.assertEqual(ftud.dssr_to_pdb_atom_id("C.C47^M"), ("C", (" ", 47, "M")))
+        self.assertEqual(ftud.dssr_to_pdb_atom_id("C.5BU47^M"), ("C",(" ", 47, "M")))
+        self.assertEqual(ftud.dssr_to_pdb_atom_id(u'A.C1'), ("A",(" ", 1, " ")))
 
 
 class TestCoaxialStacks(unittest.TestCase):